Target Editor and Fixes
Started working on a new page which is the target editor. The idea in the end is that you should be able to draw the digital targets dynamically instead of having to trace over a reference image in photoshop. This is because of the perspective correction that I have implemented that has small imperfections